+++ polars/polars-lazy/src/physical_plan/expressions/apply.rs
@@ -1,0 +1,56 @@
+use crate::physical_plan::state::ExecutionState;
+use crate::physical_plan::PhysicalAggregation;
+use crate::prelude::*;
+use polars_core::prelude::*;
+use std::sync::Arc;
+
+pub struct ApplyExpr {
+    pub input: Arc<dyn PhysicalExpr>,
+    pub function: NoEq<Arc<dyn SeriesUdf>>,
+    pub output_type: Option<DataType>,
+    pub expr: Expr,
+}
+
+impl ApplyExpr {
+    pub fn new(
+        input: Arc<dyn PhysicalExpr>,
+        function: NoEq<Arc<dyn SeriesUdf>>,
+        output_type: Option<DataType>,
+        expr: Expr,
+    ) -> Self {
+        ApplyExpr {
+            input,
+            function,
+            output_type,
+            expr,
+        }
+    }
+}
+
+impl PhysicalExpr for ApplyExpr {
+    fn as_expression(&self) -> &Expr {
+        &self.expr
+    }
+
+    fn evaluate(&self, df: &DataFrame, state: &ExecutionState) -> Result<Series> {
+        let input = self.input.evaluate(df, state)?;
+        let in_name = input.name().to_string();
+        let mut out = self.function.call_udf(input)?;
+        if in_name != out.name() {
+            out.rename(&in_name);
+        }
+        Ok(out)
+    }
+    fn to_field(&self, input_schema: &Schema) -> Result<Field> {
+        match &self.output_type {
+            Some(output_type) => {
+                let input_field = self.input.to_field(input_schema)?;
+                Ok(Field::new(input_field.name(), output_type.clone()))
+            }
+            None => self.input.to_field(input_schema),
+        }
+    }
+    fn as_agg_expr(&self) -> Result<&dyn PhysicalAggregation> {
+        Ok(self)
+    }
+}
